+++ common/routes/authtoken.js
@@ -13,7 +13,11 @@ router.get('/', function (req, res) {
   if (req.session.authToken) {
     logger.info(`Auth token REFRESH for ${req.session.passport.user}: ${req.session.authToken}`);
   } else {
-    req.session.authToken = uuid();
+    if (req.session.passport.user.endsWith('ferropoly.ch')) {
+      req.session.authToken = 'demo-user-static-auth-token';
+    } else {
+      req.session.authToken = uuid();
+    }
     logger.info(`NEW auth token for ${req.session.passport.user}: ${req.session.authToken}`);
   }